image_ai/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
google-cloud-aiplatform>=1.66.0
google-cloud-secret-manager>=2.20.0
google-auth>=2.34.0
Pillow>=10.3.0
